// fem-vision-studio/design-mockup/src/lib/types/concrete/InteractionPoint.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Eén punt van het N-M-interactiediagram (bezwijkomhullende).
 */
export type InteractionPoint = { 
/**
 * Normaalkracht in kN (trek positief).
 */
n_kn: number, 
/**
 * Momentweerstand bij die normaalkracht, kNm (≥ 0).
 */
m_rd_knm: number, };